requirements.txt
Django==1.10.1
django-common==0.1.51
djangorestframework==3.4.6
